Implemented the update method
This method only applies to users for now, but it resembles a lot to the
create method. Because of this, I've created the `createUpdate` helper
function that implements both the `create` and the `update` methods as
much as possible, and these functions have to, in turn, call this
function with the proper arguments.

I've also fixed a couple of issues on some integration tests: the
create/update endpoints return an user object, and it gets rendered
accordingly. Moreover, I've fixed the corner case in which the server
returned an empty string or a "null" string as a response.

See SUSE/Portus#1516

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>