Implemented the `vendor` flag for exec
Users can pass the `vendor` flag to the `exec` command. This flag will
instruct portusctl to use the local `vendor` directory for some of the
environment variables. One thing to note is that I tested this locally,
but I cannot provide integration tests for it for now because of the
Docker image we are currently using.

That being said, I provided a couple of integration tests for the
general functionality of the exec command.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>