Added code coverage report
This has been done by taking some ideas from openSUSE/umoci. In short,
we are merging code coverage results from unit and integration testing.
This way we get a full-fledged report. Moreover, I've added an AWK
script that will exit with a proper status depending on whether a
certain coverage threshold has been met.

That being said, I suspect that this is not entirely correct, and I plan
to fix all the corner cases as soon as possible.

Fixes #4
See #9

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>